runmodes: let thread count callback return uint16_t
It will be a long time before we need more than 64k threads.

Update capture methods.

+++ src/source-pcap.h
@@ -45,7 +45,7 @@ typedef struct PcapIfaceConfig_
 {
     char iface[PCAP_IFACE_NAME_LENGTH];
     /* number of threads */
-    int threads;
+    uint16_t threads;
     /* socket buffer size */
     int buffer_size;
     /* snapshot length */
